Support users switching accounts
Intercom('shutdown') must be called before tracking a new user or the
data will be tracked to the previous users Intercom entity. At the same
time Intercom('shutdown') can’t be called before the current tracking
has completed. To be able to wait for the previous tracking to complete
I’ve created a queue, which solves the problem.